requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
psycopg2-binary
openai
python-dotenv
requests
